Add dedicated clusterHealth store
