+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@cratedb/crate-gc-admin",
-  "version": "0.21.5",
+  "version": "0.21.6",
   "author": "cratedb",
   "private": false,
   "type": "module",
